pyproject.toml
[build-system]
requires = ["setuptools>=66", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vyt-pipe"
version = "0.1.0"
description = "Skeleton CLI pipeline for VYT-PIPE"
authors = [{ name = "VYT Team" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer[all]==0.12.5",
    "rich==13.9.2",
    "PyYAML==6.0.2",
    "Pillow==10.4.0",
    "opencv-python==4.10.0.84",
    "numpy==2.1.1",
    "scikit-image==0.24.0",
    "shapely==2.0.6",
    "svgwrite==1.4.3",
    "svgpathtools==1.6.1",
    "cairosvg==2.7.1",
    "pikepdf==9.3.2",
    "python-pptx==0.6.23"
]

[project.scripts]
vyt = "vyt.cli.__main__:app"
